requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.29.0

psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

pandas>=2.1.4
numpy>=1.26.2
faker>=20.1.0

chromadb>=0.4.22
sqlglot>=20.0.0
pydantic>=2.5.3

google-genai>=1.0.0
requests>=2.31.0
